rename workers to unify names
